[Github Search] Moving the Activities class files to the right package
