+++ GithubSearch/app/src/main/java/br/com/bernardino/githubsearch/ui/activities/SplashActivity.kt
@@ -1,8 +1,9 @@
-package br.com.bernardino.githubsearch
+package br.com.bernardino.githubsearch.ui.activities
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import br.com.bernardino.githubsearch.R
 import kotlinx.coroutines.*
 
 class SplashActivity : AppCompatActivity() {
